feat: #107 🎸 pages/rooms内のダイアログを別ファイルに切り分け
